alooooo

+++ app/views.py
@@ -1,15 +1,101 @@
 from django.shortcuts import render,get_object_or_404,redirect
 from .models import *
 
+# def header(request):
+#     user_id = request.session.get('user_id')
+#     if not user_id:
+#         return redirect('login')
+#     ans = {'user': UserModel.objects.get(id=user_id)}
+#     return render(request,'header.html',ans)
+
+# Authentication
+def register_view(request):
+    if request.method == 'POST':
+        fullname = request.POST.get('fullname')
+        email = request.POST.get('email')
+        password = request.POST.get('password')
+        if UserModel.objects.filter(email=email).first():
+            return render(request,'registration/register.html',{'error':'This email already exists!'})
+        UserModel.objects.create(
+            fullname=fullname,
+            email=email,
+            password=password
+        )
+        return redirect('login')
+        
+    return render(request, 'registration/register.html')
+
+def login_view(request):
+    if request.method == 'POST':
+        email = request.POST.get('email')
+        password = request.POST.get('password')
+
+        user = UserModel.objects.filter(email=email, password=password).first()
+
+        if user:
+            request.session['user_id'] = user.id
+            return redirect('/')
+        else:
+            return render(request, 'registration/login.html', {'error': 'Error email or password'})
+            
+    return render(request, 'registration/login.html')
+
+def logout_view(request):
+    request.session.flush()
+    return redirect('login')
+
+def forgot_view(request):
+    if request.method == 'POST':
+        email = request.POST.get('email')
+        new_pass = request.POST.get('password')
+        user = UserModel.objects.filter(email=email).first()
+        
+        if user:
+            user.password=new_pass
+            user.save()
+            return redirect('login')
+        else:
+            return render(request, 'registration/forgot_password.html', {'error': "We haven't such email"})
+            
+    return render(request, 'registration/forgot_password.html')
+
+def change_view(request):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
+    if request.method == 'POST':
+        old_pass = request.POST.get('old_pass')
+        new_pass = request.POST.get('new_pass')
+        user = UserModel.objects.filter(id=user_id,password=old_pass).first()
+        if user:
+            user.password = new_pass
+            user.save()
+            return redirect('/')
+        else:
+            return render(request,'registration/change_password.html',{'error':'Current password not true !'})
+
+    return render(request, 'registration/change_password.html')
+
+
+
 def base(request):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     return render(request,'base.html')
 
 # Category CRUD
 def categories(request):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     ans = {'categories': Category.objects.all()}
     return render(request,'categories/categories.html',ans)
 
 def categories_create(request):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     if request.method == 'POST':
         Category.objects.create(
             name = request.POST.get('name'),
@@ -19,6 +105,9 @@ def categories_create(request):
     return render(request,'categories/categ_create_or_update.html')
 
 def categories_update(request, pk):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     item = get_object_or_404(Category,id=pk)
     if request.method == 'POST':
         item.name = request.POST.get('name')
@@ -28,25 +117,40 @@ def categories_update(request, pk):
     return render(request,'categories/categ_create_or_update.html',{'category':item})
 
 def categories_delete(request, pk):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     item = get_object_or_404(Category,id=pk)
     item.delete()
     return redirect('categories')
 
 def categories_detail(request, pk):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     item = get_object_or_404(Category,id=pk)
     return render(request,'categories/category.html',{'category':item})
     
 
 # Author CRUD
 def authors(request):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     items = Author.objects.all()
     return render(request, 'authors/authors.html', {'authors': items})
 
 def authors_detail(request, pk):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     item = get_object_or_404(Author, id=pk)
     return render(request, 'authors/author.html', {'author': item})
 
 def authors_create(request):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     if request.method == 'POST':
         Author.objects.create(
             full_name=request.POST.get('full_name'),
@@ -58,6 +162,9 @@ def authors_create(request):
     return render(request, 'authors/author_create_or_update.html')
 
 def authors_update(request, pk):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     item = get_object_or_404(Author, id=pk)
     if request.method == 'POST':
         item.full_name = request.POST.get('full_name')
@@ -70,20 +177,32 @@ def authors_update(request, pk):
     return render(request, 'authors/author_create_or_update.html', {'author': item})
 
 def authors_delete(request, pk):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     item = get_object_or_404(Author, id=pk)
     item.delete()
     return redirect('authors')
 
 # Publisher CRUD
 def publishers(request):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     ans = {'publishers': Publisher.objects.all()}
     return render(request,'publishers/publishers.html',ans)
 
 def publishers_detail(request, pk):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     item = get_object_or_404(Publisher,id=pk)
     return render(request,'publishers/publisher.html',{'publisher':item})
 
 def publishers_create(request):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     if request.method == 'POST':
         Publisher.objects.create(
             name = request.POST.get('name'),
@@ -95,6 +214,9 @@ def publishers_create(request):
     return render(request,'publishers/pub_create_or_update.html')
 
 def publishers_update(request, pk):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     item = get_object_or_404(Publisher,id=pk)
     if request.method == 'POST':
         item.name = request.POST.get('name')
@@ -107,12 +229,18 @@ def publishers_update(request, pk):
     return render(request,'publishers/pub_create_or_update.html',{'publisher':item})
 
 def publishers_delete(request, pk):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     item = get_object_or_404(Publisher,id=pk)
     item.delete()
     return redirect('publishers')
 
 # Book CRUD
 def books(request):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     all_books = Book.objects.all()
     categories = Category.objects.all()
 
@@ -141,10 +269,16 @@ def books(request):
     return render(request, 'books/books.html', ans)
 
 def books_detail(request, pk):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     item = get_object_or_404(Book,id=pk)
     return render(request,'books/book.html',{'book':item})
 
 def books_create(request):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     if request.method == 'POST':
         Book.objects.create(
             title = request.POST.get('title'),
@@ -165,6 +299,9 @@ def books_create(request):
     return render(request,'books/book_create_or_update.html',choices)
 
 def books_update(request, pk):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     item = get_object_or_404(Book,id=pk)
     if request.method == 'POST':
         item.title = request.POST.get('title')
@@ -185,6 +322,9 @@ def books_update(request, pk):
     return render(request,'books/book_create_or_update.html', {'book': item, 'choices': choices})
 
 def books_delete(request, pk):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('login')
     item = get_object_or_404(Book,id=pk)
     item.delete()
     return redirect('books')
